\name{tikzDevice-package}
\Rdversion{1.1}
\alias{tikzDevice-package}
\alias{tikzDevice}
\docType{package}
\title{
Support for native LaTeX output of R graphics
}
\description{
This package implements the \code{\link{tikz}} device which allows for the output of R graphics 
functions to be saved in a LaTeX friendly format.
}
\details{
\tabular{ll}{
Package: \tab tikzDevice\cr
Type: \tab Package\cr
Version: \tab 0.35\cr
Date: \tab 2009-07-16\cr
License: \tab GPL 2.0 or greater\cr
LazyLoad: \tab yes\cr
}
}
\author{
Cameron Bracken

Charlie Sharpsteen: \email{source@sharpsteen.net}

Submit bug reports to: \email{tikzdevice-bugs@lists.r-forge.r-project.org}
}
\references{
The TikZ and PGF Packages: Manual for version 2.00\cr
\url{http://sourceforge.net/projects/pgf}\cr
Till Tantu, February 20, 2008
}
